refactor: 트랜잭션 매니저 관련 클래스 패키지 이동
